+++ src/components/StatementBrowser/AddValue/AddValueTemplate.js
@@ -30,6 +30,33 @@ export default function AddValueTemplate(props) {
     const [formFeedback, setFormFeedback] = useState(null);
     const [templateIsLoading, setTemplateIsLoading] = useState(false); // to show loading indicator of the template if the value class has a template
 
+    // uniqueLabel is set to true when it's a research problem
+    const [uniqueLabel, setuniqueLabel] = useState(props.valueClass && props.valueClass.id === process.env.REACT_APP_CLASSES_PROBLEM ? true : false);
+    const [disabledCreate, setDisabledCreate] = useState(false);
+
+    const handleCreateExistingLabel = (inputValue, selectOptions) => {
+        //check if label exists
+        if (
+            uniqueLabel &&
+            inputValue &&
+            selectOptions
+                .map(s =>
+                    String(s.label)
+                        .trim()
+                        .toLowerCase()
+                )
+                .includes(
+                    String(inputValue)
+                        .trim()
+                        .toLowerCase()
+                )
+        ) {
+            setDisabledCreate(true);
+        } else {
+            setDisabledCreate(false);
+        }
+    };
+
     useEffect(() => {
         if (valueType === 'literal' && literalInputRef.current) {
             literalInputRef.current.focus();
@@ -41,6 +68,10 @@ export default function AddValueTemplate(props) {
     useEffect(() => {
         setValueType(props.isLiteral ? 'literal' : 'object');
     }, [props.isLiteral]);
+
+    useEffect(() => {
+        setuniqueLabel(props.valueClass && props.valueClass.id === process.env.REACT_APP_CLASSES_PROBLEM ? true : false);
+    }, [props.valueClass]);
 
     useEffect(() => {
         if (!showAddValue) {
@@ -175,7 +206,8 @@ export default function AddValueTemplate(props) {
                                         props.selectResource({
                                             increaseLevel: true,
                                             resourceId: resourceId,
-                                            label: isInlineResource
+                                            label: isInlineResource,
+                                            propertyLabel: props.predicate.label
                                         });
                                     }
                                 });
@@ -238,6 +270,7 @@ export default function AddValueTemplate(props) {
                                     }
                                 }}
                                 innerRef={ref => (resourceInputRef.current = ref)}
+                                handleCreateExistingLabel={handleCreateExistingLabel}
                             />
                         ) : (
                             <InputField
@@ -272,11 +305,18 @@ export default function AddValueTemplate(props) {
                             </StyledButton>
                             <StyledButton
                                 outline
+                                disabled={!inputValue || disabledCreate}
                                 onClick={() => {
                                     onSubmit();
                                 }}
                             >
-                                Create
+                                {disabledCreate ? (
+                                    <Tippy content="Please use the existing research problem that has this label." arrow={true}>
+                                        <span>Create</span>
+                                    </Tippy>
+                                ) : (
+                                    'Create'
+                                )}
                             </StyledButton>
                         </InputGroupAddon>
                     </InputGroup>