+++ src/network.js
@@ -8,6 +8,7 @@ export const annotationServiceUrl = process.env.REACT_APP_ANNOTATION_SERVICE_URL
 export const resourcesUrl = `${url}resources/`;
 export const organizationsUrl = `${url}organizations/`;
 export const observatoriesUrl = `${url}observatories/`;
+export const problemsUrl = `${url}problems/`;
 export const predicatesUrl = `${url}predicates/`;
 export const userUrl = `${url}user/`;
 export const statementsUrl = `${url}statements/`;
@@ -368,8 +369,10 @@ export const getRDFDataCubeVocabularyClasses = () => {
     return submitGetRequest(`${classesUrl}?q=qb:`);
 };
 
-export const getAllClasses = () => {
-    return submitGetRequest(`${classesUrl}`);
+export const getAllClasses = ({ page = 1, items = 9999, sortBy = 'created_at', desc = true, q = null }) => {
+    const params = queryString.stringify({ page: page, items: items, sortBy: sortBy, desc: desc, ...(q ? { q: q } : {}) });
+
+    return submitGetRequest(`${classesUrl}?${params}`);
 };
 
 export const saveFullPaper = (data, mergeIfExists = false) => {
@@ -551,70 +554,39 @@ export const getTemplateById = templateId => {
                 }
             );
 
-            return Promise.all([components]).then(templateComponents => {
-                const subTemplates = templateStatements
-                    .filter(statement => statement.predicate.id === process.env.REACT_APP_TEMPLATE_SUB_TEMPLATE)
+            return Promise.all([components]).then(templateComponents => ({
+                id: templateId,
+                label: template.label,
+                statements: templateStatements.map(s => s.id),
+                predicate: templatePredicate
+                    ? {
+                          id: templatePredicate.object.id,
+                          label: templatePredicate.object.label
+                      }
+                    : {},
+                labelFormat: templateFormatLabel ? templateFormatLabel.object.label : '',
+                hasLabelFormat: templateFormatLabel ? true : false,
+                isStrict: templateIsStrict ? true : false,
+                components: templateComponents[0].sort((c1, c2) => sortMethod(c1.order, c2.order)),
+                class: templateClass
+                    ? {
+                          id: templateClass.object.id,
+                          label: templateClass.object.label
+                      }
+                    : {},
+                researchFields: templateStatements
+                    .filter(statement => statement.predicate.id === process.env.REACT_APP_TEMPLATE_OF_RESEARCH_FIELD)
                     .map(statement => ({
                         id: statement.object.id,
                         label: statement.object.label
-                    }));
-                return Promise.all(
-                    subTemplates.map(template =>
-                        getStatementsBySubject({ id: template.id }).then(subTemplateStatements => {
-                            const subTemplatePredicate = subTemplateStatements
-                                .filter(statement => statement.predicate.id === process.env.REACT_APP_TEMPLATE_OF_PREDICATE)
-                                .map(statement => ({
-                                    id: statement.object.id,
-                                    label: statement.object.label
-                                }));
-                            const subTemplateClass = subTemplateStatements
-                                .filter(statement => statement.predicate.id === process.env.REACT_APP_TEMPLATE_OF_CLASS)
-                                .map(statement => ({
-                                    id: statement.object.id,
-                                    label: statement.object.label
-                                }));
-                            return {
-                                ...template,
-                                predicate: subTemplatePredicate[0],
-                                class: subTemplateClass && subTemplateClass.length > 0 ? subTemplateClass[0] : null
-                            };
-                        })
-                    )
-                ).then(subs => ({
-                    id: templateId,
-                    label: template.label,
-                    statements: templateStatements.map(s => s.id),
-                    predicate: templatePredicate
-                        ? {
-                              id: templatePredicate.object.id,
-                              label: templatePredicate.object.label
-                          }
-                        : {},
-                    labelFormat: templateFormatLabel ? templateFormatLabel.object.label : '',
-                    hasLabelFormat: templateFormatLabel ? true : false,
-                    isStrict: templateIsStrict ? true : false,
-                    components: templateComponents[0].sort((c1, c2) => sortMethod(c1.order, c2.order)),
-                    class: templateClass
-                        ? {
-                              id: templateClass.object.id,
-                              label: templateClass.object.label
-                          }
-                        : {},
-                    researchFields: templateStatements
-                        .filter(statement => statement.predicate.id === process.env.REACT_APP_TEMPLATE_OF_RESEARCH_FIELD)
-                        .map(statement => ({
-                            id: statement.object.id,
-                            label: statement.object.label
-                        })),
-                    researchProblems: templateStatements
-                        .filter(statement => statement.predicate.id === process.env.REACT_APP_TEMPLATE_OF_RESEARCH_PROBLEM)
-                        .map(statement => ({
-                            id: statement.object.id,
-                            label: statement.object.label
-                        })),
-                    subTemplates: subs
-                }));
-            });
+                    })),
+                researchProblems: templateStatements
+                    .filter(statement => statement.predicate.id === process.env.REACT_APP_TEMPLATE_OF_RESEARCH_PROBLEM)
+                    .map(statement => ({
+                        id: statement.object.id,
+                        label: statement.object.label
+                    }))
+            }));
         })
     );
 };
@@ -665,15 +637,15 @@ export const getOrganization = id => {
     return submitGetRequest(`${organizationsUrl}${encodeURIComponent(id)}/`);
 };
 
-export const createOrganization = (organizationName, organizationLogo, createdBy) => {
-    return submitPostRequest(organizationsUrl, { 'Content-Type': 'application/json' }, { organizationName, organizationLogo, createdBy });
+export const createOrganization = (organizationName, organizationLogo, createdBy, url) => {
+    return submitPostRequest(organizationsUrl, { 'Content-Type': 'application/json' }, { organizationName, organizationLogo, createdBy, url });
 };
 
-export const getAllObservatoriesbyOrganizationId = id => {
+export const getAllObservatoriesByOrganizationId = id => {
     return submitGetRequest(`${organizationsUrl}${encodeURIComponent(id)}/observatories`);
 };
 
-export const getObservatorybyId = id => {
+export const getObservatoryById = id => {
     return submitGetRequest(`${observatoriesUrl}${encodeURIComponent(id)}/`);
 };
 
@@ -681,12 +653,24 @@ export const getUsersByObservatoryId = id => {
     return submitGetRequest(`${observatoriesUrl}${encodeURIComponent(id)}/users`);
 };
 
-export const getResourcesByObservatoryId = id => {
-    return submitGetRequest(`${observatoriesUrl}${encodeURIComponent(id)}/resources`);
+export const getUsersByOrganizationId = id => {
+    return submitGetRequest(`${organizationsUrl}${encodeURIComponent(id)}/users`);
 };
 
-export const createObservatory = (observatoryName, organizationId) => {
-    return submitPostRequest(observatoriesUrl, { 'Content-Type': 'application/json' }, { observatoryName, organizationId });
+export const getResourcesByObservatoryId = id => {
+    return submitGetRequest(`${observatoriesUrl}${encodeURIComponent(id)}/papers`);
+};
+
+export const getComparisonsByObservatoryId = id => {
+    return submitGetRequest(`${observatoriesUrl}${encodeURIComponent(id)}/comparisons`);
+};
+
+export const getProblemsByObservatoryId = id => {
+    return submitGetRequest(`${observatoriesUrl}${encodeURIComponent(id)}/problems`);
+};
+
+export const createObservatory = (observatoryName, organizationId, description) => {
+    return submitPostRequest(observatoriesUrl, { 'Content-Type': 'application/json' }, { observatoryName, organizationId, description });
 };
 
 export const getContributorsByResourceId = id => {
@@ -704,17 +688,21 @@ export const getContributorsByResourceId = id => {
 };
 
 export const getObservatoryAndOrganizationInformation = (observatoryId, organizationId) => {
-    return getObservatorybyId(observatoryId).then(obsResponse => {
+    return getObservatoryById(observatoryId).then(obsResponse => {
         return getOrganization(organizationId).then(orgResponse => {
             return {
                 id: observatoryId,
                 name: obsResponse.name.toUpperCase(),
                 organization: {
                     id: organizationId,
-                    name: orgResponse.organization_name,
-                    logo: orgResponse.organization_logo
+                    name: orgResponse.name,
+                    logo: orgResponse.logo
                 }
             };
         });
     });
+};
+
+export const getResearchFieldsByResearchProblemId = problemId => {
+    return submitGetRequest(`${problemsUrl}${encodeURIComponent(problemId)}/fields`);
 };